+++ include/core/lethe_grid_tools.h
@@ -402,12 +402,51 @@ namespace LetheGridTools
   };
 
   /**
+   * @brief Transform angular velocity into linear velocity considering that the
+   * angular velocity @p omega is constant in all directions.
+   *
+   * @param[in] omega Angular velocity value
+   * @param[in] point Coordinates of the point at which the linear velocity is
+   * being calculated
+   * @param[in] rotation_axis Rotation axis used in 3D cases, so that the
+   * angular velocity vector is given by omega_vec = omega * rotation_axis. In
+   * 2D, we apply a rotation around the z-axis.
+   *
+   * @return Linear velocity at @p point
+   */
+  template <int dim>
+  inline Tensor<1, dim>
+  angular_to_linear_velocity(
+    const double          omega,
+    const Tensor<1, dim> &point,
+    const Tensor<1, dim> &rotation_axis = Tensor<1, dim>())
+  {
+    Tensor<1, dim> result;
+
+    if constexpr (dim == 2)
+      {
+        result[0] = -omega * point[1];
+        result[1] = omega * point[0];
+      }
+
+    if constexpr (dim == 3)
+      {
+        // Store angular velocity according to unit vector that defines the
+        // rotation axis
+        const auto omega_vec = omega * rotation_axis;
+        // Compute cross product
+        result = cross_product_3d(omega_vec, point);
+      }
+
+    return result;
+  }
+
+  /**
    * @brief Rotate the mapping according to the rotation angle in a rotor-stator configuration
    *
    * @param[in] dof_handler DofHandler of the triangulation
    * @param[in,out] mapping_cache Rotated mapping
    * @param[in] mapping Current mapping
-   * @param[in] radius Radius or the rotor domain which is going to be rotated
    * @param[in] rotation_angle Rotation angle of the rotor domain
    * @param[in] center_of_rotation Center of rotation of the rotor domain.
    * Default is the origin
@@ -418,7 +457,6 @@ namespace LetheGridTools
   rotate_mapping(const DoFHandler<dim> &dof_handler,
                  MappingQCache<dim>    &mapping_cache,
                  const Mapping<dim>    &mapping,
-                 const double          &radius,
                  const double          &rotation_angle,
                  const Point<dim>      &center_of_rotation = Point<dim>(),
                  const Tensor<1, dim>  &rotation_axis      = Tensor<1, dim>());