+++ source/solvers/fluid_dynamics_matrix_based.cc
@@ -865,8 +865,7 @@ FluidDynamicsMatrixBased<dim>::assemble_local_system_matrix(
 
   if (this->simulation_parameters.mortar_parameters.enable)
     scratch_data.reinit_mortar(cell,
-                               this->simulation_parameters.mortar_parameters,
-                               this->mortar_manager->radius[0]);
+                               this->simulation_parameters.mortar_parameters);
 
   scratch_data.calculate_physical_properties();
 
@@ -1105,8 +1104,7 @@ FluidDynamicsMatrixBased<dim>::assemble_local_system_rhs(
 
   if (this->simulation_parameters.mortar_parameters.enable)
     scratch_data.reinit_mortar(cell,
-                               this->simulation_parameters.mortar_parameters,
-                               this->mortar_manager->radius[0]);
+                               this->simulation_parameters.mortar_parameters);
 
   scratch_data.calculate_physical_properties();