+++ source/solvers/navier_stokes_scratch_data.cc
@@ -410,38 +410,40 @@ template <int dim>
 void
 NavierStokesScratchData<dim>::reinit_mortar(
   const typename DoFHandler<dim>::active_cell_iterator &cell,
-  const Parameters::Mortar<dim>                        &mortar_parameters,
-  const double                                         &radius)
+  const Parameters::Mortar<dim>                        &mortar_parameters)
 {
-  const auto cell_center = cell->center();
-
   // Get updated rotor angular velocity
   mortar_parameters.rotor_angular_velocity->set_time(
     this->simulation_control->get_current_time());
   const double rotor_angular_velocity =
     mortar_parameters.rotor_angular_velocity->value(Point<dim>());
 
-  // Compute radius between center of rotation and current cell center
-  const double radius_current =
-    cell_center.distance(mortar_parameters.center_of_rotation);
+  // Apply prescribed rotor angular velocity only at rotor cells
+  // (material_id = 1)
+  double omega = 0.0;
+  if (cell->material_id() == 1)
+    omega = rotor_angular_velocity;
 
-  // Use prescribed rotor angular velocity only if cell is part of the rotor
-  double cell_rotor_angular_velocity;
-  if (radius_current > radius)
-    cell_rotor_angular_velocity = 0.0;
-  else
-    cell_rotor_angular_velocity = rotor_angular_velocity;
+  // The mortar implementation is not defined for more than two
+  // materials. Throw when running debug mode as a security measure to prevent
+  // this case
+  Assert(
+    cell->material_id() < 2,
+    ExcMessage(
+      "The material id in a cell was identified as equal or greater than 2, however the mortar implementation is not defined for more than two materials."));
 
   // Compute rotor linear velocity at quadrature points
   rotor_linear_velocity_values = std::vector<Tensor<1, dim>>(this->n_q_points);
   for (unsigned int q = 0; q < n_q_points; ++q)
     {
-      // Assumption in 3D case: rotation axis is in z
-      // TODO generalize rotation axis
-      const auto x                       = fe_values.quadrature_point(q)[0];
-      const auto y                       = fe_values.quadrature_point(q)[1];
-      rotor_linear_velocity_values[q][0] = -cell_rotor_angular_velocity * y;
-      rotor_linear_velocity_values[q][1] = cell_rotor_angular_velocity * x;
+      // Shift quadrature point by the center of rotation
+      const auto p =
+        fe_values.quadrature_point(q) - mortar_parameters.center_of_rotation;
+
+      // Compute terms of u_ale
+      rotor_linear_velocity_values[q] =
+        LetheGridTools::angular_to_linear_velocity(
+          omega, p, mortar_parameters.rotation_axis);
 
       // Update velocity for stabilization
       this->velocity_for_stabilization[q] -= rotor_linear_velocity_values[q];